fix(nc-gui): apply column color on url filter
